+++ src/fl/five_bit_hd_gamma.h
@@ -4,15 +4,11 @@
 #pragma once
 
 #include "fl/gamma.h"
-#include "fl/stdint.h"
 #include "fl/int.h"
+#include "fl/math.h"
 
 #include "crgb.h"
 #include "lib8tion/scale8.h"
-#include "fl/force_inline.h"
-#include "fl/math.h"
-#include "fl/namespace.h"
-#include "lib8tion/brightness_bitshifter.h"
 
 namespace fl {
 
@@ -47,9 +43,8 @@ void internal_builtin_five_bit_hd_gamma_bitshift(CRGB colors, CRGB colors_scale,
                                                  fl::u8 *out_power_5bit);
 
 // Exposed for testing.
-fl::u8 five_bit_bitshift(u16 r16, u16 g16, u16 b16,
-                          fl::u8 brightness, CRGB *out,
-                          fl::u8 *out_power_5bit);
+void five_bit_bitshift(u16 r16, u16 g16, u16 b16, fl::u8 brightness, CRGB *out,
+                       fl::u8 *out_power_5bit);
 
 #ifdef FASTLED_FIVE_BIT_HD_BITSHIFT_FUNCTION_OVERRIDE
 // This function is located somewhere else in your project, so it's declared
@@ -124,65 +119,73 @@ inline void internal_builtin_five_bit_hd_gamma_bitshift(
                       out_power_5bit);
 }
 
-inline fl::u8 five_bit_bitshift(u16 r16, u16 g16, u16 b16,
-                                 fl::u8 brightness, CRGB *out,
-                                 fl::u8 *out_power_5bit) {
+// Since the return value wasn't used, it has been omitted.
+// It's not clear what scale brightness is, or how it is to be applied,
+// so we assume 8 bits applied over the given rgb values.
+inline void five_bit_bitshift(uint16_t r16, uint16_t g16, uint16_t b16,
+                              uint8_t brightness, CRGB *out,
+                              uint8_t *out_power_5bit) {
 
-    auto max3 = [](u16 a, u16 b, u16 c) {
-        return fl_max(fl_max(a, b), c);
-    };
+    // NEW in 3.10.2: A new closed form solution has been found!
+    // Thank you https://github.com/gwgill!
+    // It's okay if you don't know how this works, few do, but it tests
+    // very well and is better than the old iterative approach which had
+    // bad quantization issues (sudden jumps in brightness in certain intervals).
+
+    // ix/31 * 255/65536 * 256 scaling factors, valid for indexes 1..31
+    static uint32_t bright_scale[32] = {
+        0,      2023680, 1011840, 674560, 505920, 404736, 337280, 289097,
+        252960, 224853,  202368,  183971, 168640, 155668, 144549, 134912,
+        126480, 119040,  112427,  106509, 101184, 96366,  91985,  87986,
+        84320,  80947,   77834,   74951,  72274,  69782,  67456,  65280};
+
+    auto max3 = [](u16 a, u16 b, u16 c) { return fl_max(fl_max(a, b), c); };
+
 
     if (brightness == 0) {
         *out = CRGB(0, 0, 0);
         *out_power_5bit = 0;
-        return 0;
+        return;
     }
     if (r16 == 0 && g16 == 0 && b16 == 0) {
         *out = CRGB(0, 0, 0);
         *out_power_5bit = (brightness <= 31) ? brightness : 31;
-        return brightness;
+        return;
     }
 
-    // Note: One day someone smarter than me will come along and invent a closed
-    // form solution for this. However, the numerical method works extremely
-    // well and has been optimized to avoid division performance penalties as
-    // much as possible.
+    uint8_t r8 = 0, g8 = 0, b8 = 0;
 
-    // Step 1: Initialize brightness
-    static const fl::u8 kStartBrightness = 0b00010000;
-    fl::u8 v5 = kStartBrightness;
-    // Step 2: Boost brightness by swapping power with the driver brightness.
-    brightness_bitshifter8(&v5, &brightness, 4);
-
-    // Step 3: Boost brightness of the color channels by swapping power with the
-    // driver brightness.
-    u16 max_component = max3(r16, g16, b16);
-    // five_bit_color_bitshift(&r16, &g16, &b16, &v5);
-    fl::u8 shifts = brightness_bitshifter16(&v5, reinterpret_cast<fl::u16*>(&max_component), 4, 2);
-    if (shifts) {
-        r16 = r16 << shifts;
-        g16 = g16 << shifts;
-        b16 = b16 << shifts;
-    }
-
-    // Step 4: scale by final brightness factor.
+    // Apply any brightness setting (we assume brightness is 0..255)
     if (brightness != 0xff) {
         r16 = scale16by8(r16, brightness);
         g16 = scale16by8(g16, brightness);
         b16 = scale16by8(b16, brightness);
     }
 
-    // brighten hardware brightness by turning on low order bits
-    if (v5 > 1) {
-        // since v5 is a power of two, subtracting one will invert the leading
-        // bit and invert all the bits below it. Example: 0b00010000 -1 =
-        // 0b00001111 So 0b00010000 | 0b00001111 = 0b00011111
-        v5 = v5 | (v5 - 1);
+    // Locate the largest value to set the brightness/scale factor
+    uint16_t scale = max3(r16, g16, b16);
+
+    if (scale == 0) {
+        *out = CRGB(0, 0, 0);
+        *out_power_5bit = 0;
+        return;
+    } else {
+        uint32_t scalef;
+
+        // Compute 5 bit quantized scale that is at or above the maximum value.
+        scale = (scale + (2047 - (scale >> 5))) >> 11;
+
+        // Adjust the 16 bit values to account for the scale, then round to 8
+        // bits
+        scalef = bright_scale[scale];
+        r8 = (r16 * scalef + 0x808000) >> 24;
+        g8 = (g16 * scalef + 0x808000) >> 24;
+        b8 = (b16 * scalef + 0x808000) >> 24;
+
+        *out = CRGB(r8, g8, b8);
+        *out_power_5bit = scale;
+        return;
     }
-    // Step 5: Convert back to 8-bit and output.
-    *out = CRGB(map16_to_8(r16), map16_to_8(g16), map16_to_8(b16));
-    *out_power_5bit = v5;
-    return brightness;
 }
 
 } // namespace fl